Setup card instantiation and collection
Refactored cards so they are still easy to create from editor and should be able to be created for basic stat editing ones without any additional coding, cards are deleted on collection and applied, probably need to store them so they can be removed if unequipped. Also need to refactor playercontroller so it actually uses the stats from EntityStats rather than stats in its own script.